Parse and process quantity annotations on definitions, let- and pi-binders and some simple tests (#6)
* Added optional quantities to let and pi in Core parser.

* Fix error message for LinearMisuse

* Fix initial quantity

* Add simple test cases for the quantity/linearity checker

* Add relevance annotations to definitions

Note: diverging from Idris2, runtime-relevant definitions are assigned quantity `Rig1`, not `RigW`. This is closer in line with QTT.

The upside is that these are correct when calling `check`/`linearCheck` (see `Core.Syntax.Decls.processDef`).

The downside is that we need to squash (see `Core.TT.RigCount.relevance`) `check`'s quantity scaling factor back to just its relevance when checking whether a definition is safe to use. The squashing is necessary because `check` combines runtime-relevance (\in {0,1}) and the quantity scaling factor (\in {0,1,w}) introduced by, e.g., function applications.

* Add relevance test

* Fix initial quantity in processHNF