Finish AutoSearch and add tests
Needed a tweak in unifySpine - in Idris 2, the spine consists of
unevaluated closures, so if we solve a metavariable in one argument, the
solution gets used when evaluating the next. Here, the spine consists of
Glued values, already evaluated, so we need to evaluate it again with
the new knowledge from unifying earlier arguments. In theory quote is
cheap, so this should be fine!