[ bug ] progress stopped on a unification issue
The hole ?A has a type that ends in:

```
(%pi Explicit Nothing
    (Main.ValidEvalContextView.View ((Main.Subset.Element Main.EvalContext.Nil) p))
    (Main.ValidEvalContextView.View ((Main.Subset.Element Main.EvalContext.Nil) p))))))
```

That's the type of the let-bound expression `eq` and the goal.
They match exactly.

Trying to feed `eq` into `?A` however gives us:

```
Krivine:508:3--511:3:When elaborating right hand side of Main.ValidEvalContextView.view:
Krivine:510:10--511:3:?Main.ValidEvalContextView.{a:4352}_[(0, b[3]), (, p[2]), (, p[2])] and b[3] are not equal
```