replace/case fixes
We record whether a 'case' arises from a top level pattern match or a
case block, because then we know whether we need to block reduction of
an application or not. This is to get the same behaviour on top level
pattern matching definitions as Idris 2, and needed because we sometimes
need to reduce under (programmer-written) case blocks.

Also in replace, we block reduction under implementations of interfaces,
because they are a special case that can expand forever if unchecked.