Better propagation of forced equalities
Instead of maintaining a big list of all the things an argument is equal
to, which gets big, keep the canonical form of each argument by updating
variables according to forced equalities from the case tree. Then we
only need to check canonical forms when building the size change graph.