Use a 'Glued' as argument to VLam
That is, don't wrap it in Core. This is what we used to do, but changed
so that we didn't evaluate arguments to binders unnecessarily. However,
that's only really an issue when unelaborating pi types. Otherwise, it
is a needless overhead.
This turns out to give a noticeable but small improvement when
evaluating large expressions.