Add a HolesOnly flag to nf
We sort of had this but it didn't behave the way I wanted it too because
quoting doesn't go under delays. We need it for the unelaboration step
in records, otherwise it might add a new top level hole rather than the
solution of the metavariable when generating types for getters.